Refactor: Project structure reorganization and initial setup
Use Vite to reorganize project structure make it more dev friendly.
